+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 * Default `initOnInsert` to true
 * Removed IE8 "support"
+* Removed `Mutiny.util.isArray` and `Mutiny.util.isString` helper methods
 
 #### [v1.2.0 (2015-08-20)](https://github.com/enova/mutiny/compare/v1.1.1...v1.2.0)